削除機能ユーザーのみ設定

+++ app/controllers/items_controller.rb
@@ -35,10 +35,14 @@ class ItemsController < ApplicationController
   end
 
   def destroy
-    if @item.destroy
-      redirect_to root_path
+    if current_user.id == @item.user_id
+      if @item.destroy
+        redirect_to root_path
+      else
+        render :show, status: :unprocessable_entity
+      end
     else
-      render :show, status: :unprocessable_entity
+      redirect_to root_path
     end
   end
